cistern: address comment
