+++ cistern/pump.go
@@ -20,7 +20,7 @@ import (
 	pb "github.com/pingcap/tipb/go-binlog"
 )
 
-const defaultBinlogChanSize int64 = 16 << 12
+const defaultBinlogChanLength int64 = 16 << 12
 
 type binlogEntity struct {
 	tp       pb.BinlogType
@@ -75,7 +75,7 @@ func NewPump(nodeID string, clusterID uint64, host string, timeout time.Duration
 		tiStore:    tiStore,
 		window:     w,
 		timeout:    timeout,
-		binlogChan: make(chan *binlogEntity, defaultBinlogChanSize),
+		binlogChan: make(chan *binlogEntity, defaultBinlogChanLength),
 	}, nil
 }
